renaming the android package name from com.reactlibrary to io.linkpreview
